Lots of changes here. Refactored a lot of operations into the apiproxy package. Eventually I'll get rid of the apiinfo package and move that functionality into the apiproxy package.
Added the devportal package which just controls creating devportal
entries. Currently this only happens when a transport occurs.

The actualy transport of an APIProxy got moved into the apiproxy
package. Anything that uses the APIs for API Manager should be in the
apiproxy package. The transport package now just has the handler and a
wrapper around the apiproxy transport funciton to do all of the
locking, and github sync.

Finally added the metadata to openapi functionality.